Editing and sharing functions restricted for poll owner

+++ user.php
@@ -119,17 +119,24 @@
 						<?=$row['number_of_answers']?>
 					</td>
 					<!-- SHARE POLL -->
-					<td><a href="javascript: sharePoll(<?=$row['id']?>)" title="Share this poll">
-						<span style="color: green" class="glyphicon glyphicon-share" aria-hidden="true"></span>
-					</a></td>
+					<?php 
+						if($username == $_SESSION['myname']) { ?>
+						<td><a href="javascript: sharePoll(<?=$row['id']?>)" title="Share this poll">
+							<span style="color: green" class="glyphicon glyphicon-share" aria-hidden="true"></span>
+						</a></td>
+					<?php } else { ?>
+						<td>
+							<span style="color: black" title="Sharing is not available for the current user" class="glyphicon glyphicon-share" aria-hidden="true"></span>
+						</td>
+					<?php } ?>
 					<!-- ANSWER POLL -->
 					<?php if(($title == 'Polls I Answered') || checkIfUserAnswered(array('db' => $db, 'user_id' => $_SESSION['myid'], 'id' => $row['id']))) { ?>
 						<td>
-							<span style="color: black;" title="You have already answered this poll" class="glyphicon glyphicon-edit" aria-hidden="true"></span>
+							<span style="color: black" title="You have already answered this poll" class="glyphicon glyphicon-edit" aria-hidden="true"></span>
 						</td>
 					<?php } else { ?>
 						<td><a href="answer_poll.php?id=<?=$row['id']?>" title="Answer this poll">
-							<span style="color: green;" class="glyphicon glyphicon-edit" aria-hidden="true"></span>
+							<span style="color: green" class="glyphicon glyphicon-edit" aria-hidden="true"></span>
 						</a></td>
 					<?php } ?>
 					<!-- VIEW POLL DETAILS -->
@@ -141,19 +148,30 @@
 						<a href="view_poll_detailed.php?id=<?=$row['id']?>"><span style="color: blue" class="glyphicon glyphicon-stats" aria-hidden="true"></span></a>
 					</td>
 					<!-- EDIT POLL -->
-					<?php if($row['number_of_answers'] > 0) { ?>
-						<td>
-							<span style="color: black" title="This poll has answers and can no longer be edited" class="glyphicon glyphicon-cog disabled" aria-hidden="true"></span>
-						</td>
-					<?php } else { ?>
-						<td><a href="edit_poll.php?id=<?=$row['id']?>" title="Edit Poll">
-							<span style="color: blue" class="glyphicon glyphicon-cog" aria-hidden="true"></span>
+					<?php
+					if($username == $_SESSION['myname']) {
+						if($row['number_of_answers'] > 0) { ?>
+							<td>
+								<span style="color: black" title="This poll has answers and can no longer be edited" class="glyphicon glyphicon-cog disabled" aria-hidden="true"></span>
+							</td>
+						<?php } else { ?>
+							<td><a href="edit_poll.php?id=<?=$row['id']?>" title="Edit Poll">
+								<span style="color: blue" class="glyphicon glyphicon-cog" aria-hidden="true"></span>
+							</a></td>
+						<?php } ?>
+						<!-- DELETE POLL -->
+						<td><a href="javascript: confirmDelete(<?=$row['id']?>)" title="Delete Poll">
+							<span style="color: red" class="glyphicon glyphicon-trash" aria-hidden="true"></span>
 						</a></td>
+					<?php } else { ?>
+						<td>
+							<span style="color: black" title="Editing is not available for the current user" class="glyphicon glyphicon-cog disabled" aria-hidden="true"></span>
+						</td>
+						<td>
+							<span style="color: black" title="Deleting is not available for the current user" class="glyphicon glyphicon-trash disabled" aria-hidden="true"></span>
+						</td>
 					<?php } ?>
-					<!-- CELETE POLL -->
-					<td><a href="javascript: confirmDelete(<?=$row['id']?>)" title="Delete Poll">
-						<span style="color: red" class="glyphicon glyphicon-trash" aria-hidden="true"></span>
-					</a></td>
+					
 				</tr>
 							
 			<?php 	} ?>
